move remaining tests to their own files
